DEV: Extract mcp core tools
